# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
        "/api/Directory": {},
        "/api/organizations/building/{buildingId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организации по ID здания",
                "operationId": "6f60a590fcbb792975e275b750dfca75",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "organizations": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/activity/{activityId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организации по ID вида деятельности",
                "operationId": "0f30517ce479f70cd9308f39afcc18ac",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "organizations": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/radius": {
            "post": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организации в радиусе",
                "operationId": "ee0f135d977adcce967edf5657180b28",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "radius",
                                    "latitude",
                                    "longitude"
                                ],
                                "properties": {
                                    "radius": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 55.7558
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 37.6176
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "organizations": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/{organizationId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить организацию по ID",
                "operationId": "c23138ac4d07c768f6624c8945040813",
                "parameters": [
                    {
                        "name": "organizationId",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об организации"
                    },
                    "404": {
                        "description": "Организация не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Organization not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/activity/search/{activityId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организаций по виду деятельности (включая дочерние)",
                "operationId": "435fb83d10ddcfe2b0f6d9140529f308",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "organizations": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/search": {
            "post": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организаций по названию",
                "operationId": "bcdc7a10576ea8ce27b2e153a86a3308",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Название организации",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "organizations": {}
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Organizations",
            "description": "Organizations"
        }
    ]
}